requirements.txt
customtkinter>=5.2.0
pypdf>=3.0.0
tkinterdnd2>=0.4.0
Pillow>=10.0.0
pyinstaller>=5.13.0
